requirements.txt
python-dotenv==1.0.1
fastapi==0.115.8
pydantic==2.10.6
transformers==4.48.2
faiss-cpu==1.10.0
mysql-connector-python==9.2.0
uvicorn==0.34.0
torch==2.6.0
spacy==3.8.4
redis==5.2.1
